+++ app/src/main/java/cn/wthee/pcrtool/ui/tool/TweetList.kt
@@ -11,21 +11,17 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.flowWithLifecycle
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadState
-import androidx.paging.compose.collectAsLazyPagingItems
+import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.items
 import cn.wthee.pcrtool.R
 import cn.wthee.pcrtool.data.db.entity.TweetData
@@ -37,7 +33,6 @@ import cn.wthee.pcrtool.ui.compose.*
 import cn.wthee.pcrtool.ui.theme.Dimen
 import cn.wthee.pcrtool.utils.Constants
 import cn.wthee.pcrtool.utils.openWebView
-import cn.wthee.pcrtool.viewmodel.TweetViewModel
 import coil.annotation.ExperimentalCoilApi
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
@@ -50,7 +45,6 @@ import kotlinx.coroutines.launch
 
 /**
  * 推特列表
- * fixme 跳转后，回到顶部
  */
 @ExperimentalCoilApi
 @ExperimentalPagerApi
@@ -59,25 +53,14 @@ import kotlinx.coroutines.launch
 @ExperimentalAnimationApi
 @Composable
 fun TweetList(
+    tweet: LazyPagingItems<TweetData>?,
     scrollState: LazyListState,
     toDetail: (String) -> Unit,
     toComic: (Int) -> Unit,
-    tweetViewModel: TweetViewModel = hiltViewModel()
 ) {
-    val lifecycle = LocalLifecycleOwner.current.lifecycle
     val coroutineScope = rememberCoroutineScope()
 
-    tweetViewModel.getTweet()
-    val flow = tweetViewModel.tweetPageList
-    val tweet = remember(flow, lifecycle) {
-        flow?.flowWithLifecycle(lifecycle = lifecycle)
-    }?.collectAsLazyPagingItems()
-
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-    ) {
+    Box(modifier = Modifier.fillMaxSize()) {
         val visible = tweet != null && tweet.itemCount > 0
         FadeAnimation(visible = visible) {
             LazyColumn(state = scrollState) {
